requirements.txt
av==12.1.0
cubvh==0.1.0
diff_gaussian_rasterization==0.0.0
diffusers==0.24.0
einops==0.8.0
huggingface_hub==0.23.4
human_body_prior==2.2.2.0
imageio==2.34.1
jaxtyping==0.2.34
loguru==0.7.2
matplotlib==3.9.2
mediapipe==0.10.14
numpy==2.1.2
nvdiffrast==0.3.1
opencv_contrib_python==4.10.0.82
opencv_python==4.10.0.82
opencv_python_headless==4.10.0.82
Pillow==10.4.0
plyfile==1.1
pymeshlab==2023.12.post1
pyrallis==0.3.1
pyrender==0.1.45
safetensors==0.4.5
scikit_learn==1.5.2
scipy==1.14.1
torch==2.1.0
torchvision==0.16.0
tqdm==4.66.4
transformers==4.26.1
xatlas==0.0.9
xformers==0.0.22.post7+cu118